[VPlan] Unify calcCharacteristicType for both function cloning and CG.
The characteristic type is type of the call or function body mask.
Currently, we have two different versions of characteristics type
calculation, in function cloning and in CG. In this change I refactored
the code so both variants call one underlying templatized function.
Also fixed a mistake with processing pointer data types.

Change-Id: I5656ab868118f6d579293b1ce26631f443871b71
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/266717
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Tested-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>