[SYCL] Fix crash when kernel argument is a multi-dimensional array. (#2341)
This patch fixes crash due to incorrect InitializedEntity for
multi-dimensional arrays. When generating the InitializedEntity
for an element, it is necessary to descend the array. For example,
the initialized entity for s.array[x][y][z] is constructed using
initialized entities for s.array[x][y], s.array[x] and s.array.
Prior to this patch, the 'descending' was not done.

Patch by: Rajiv Deodhar and Elizabeth Andrews

Change-Id: I99b51dc16ff7884bd8221a77349185d4ebf33113
Signed-off-by: Elizabeth Andrews <elizabeth.andrews@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/266824
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>