[VPlan][Predicator] Incomplete workaround for a uniform phi in linearized control flow
This in an incomplete quick hack to solve a subpart of a more general problem
when a uniform instruction has divergent (in general) operands.

That might happen when that operand is divergent in general but happens to be
uniform on some particular control flow and the user instruction there is
uniform.

So far, known appearances are coming from the uniform phis inside linearized
control flow resulting in divergent blends (blends are unmasked so the
uniformity can't be simply propagated).

Jira-Id: CMPLRLLVM-21787

First iteration was reviewed on
https://git-amr-2.devtools.intel.com/gerrit/266957 but had to be reverted due
some stability issues.

Change-Id: I3d46d7c8efb4748b6040da53150746441d136020
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/270257
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>