Bypass mask handling for VPlan-masked function vector variants
HandleVPlanMask pass, TableGen and OCLBuiltin are
involved in this change.

  - [TableGen] Append an extra VPlan-masked
    function list to "VectInfo.gen", guarded
    by macro IMPORT_VPLAN_MASKED_VARIANTS.

  - [HandleVPlanMask] Bypass VPlan mask handling
    immediately if the function has VPlan masks.

  - [OCLBuiltin] Implement VPlan-masked acos
    builtin.

The element type of a VPlan-styled mask is the same
as the characteristic data type defined in Intel
Vector ABI document:
https://software.intel.com/content/dam/develop/external/us/en/documents/intel-vector-function-abi.pdf

While the element type of a Volcano mask is forced
to be i32 type. HandleVPlanMask was used to convert
VPlan-masked calls into Volcano-masked calls, thus
compatibilizing masked builtins using i32 masks.

Since we are shifting from Volcano to VPlan, we
should also upgrade builtins to use VPlan masks.
Therefore, HandleVPlanMask would be deprecated
when we deprecate Volcano totally.

Now we add extra controlling logic to check whether
a builtin function uses VPlan masks, if so, we
don't need to insert extra cast calls to convert
mask type.

Change-Id: I0524cff28126fec7f7d84726b3430924c67e21b6
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/267221
Reviewed-by: Yilong Guo <yilong.guo@intel.com>
Tested-by: Yilong Guo <yilong.guo@intel.com>
Reviewed-by: Sotkin, Alexey <alexey.sotkin@intel.com>