[VecClone] Remove insertInstruction().
CMPLRLLVM-11907: The problem was that the entry block already had a store
instruction. insertInstruction() groups instructions of the same opcode
together. In case of store, it was emitting all the store instructions at the
top of the entry block. But, the allocas were emitted after the store
instructions.

Hence, I removed insertInstructions. Allocas are emitted at the front of the
entry block. All the rest (stores, bitcasts, masks) are emitted at the bottom of
the entry block.

Change-Id: I2bf859915b7477d9f61163068941493e256c2a8d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/267010
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Mitropoulou, Konstantina <konstantina.mitropoulou@intel.com>
Reviewed-by: Sotkin, Alexey <alexey.sotkin@intel.com>