[InstCombine] Convert splat pointer for gather to GEP. Teach isAllocSiteRemovable to handle scatter.
An alloca that is only stored to and never loaded or
dereferenced, etc can be removed along with the stores.
We already did this for some cases, but we didn't recognize
scatter as a store.

isAllocSiteRemovable doesn't understand the splat form so
convert it to a GEP with 0 vector index. This is better
for isel anyway.

This helps with the specific case cited in CMPLRLLVM-22157
though the vectorizer should still address that if possible.

Change-Id: Iad3d628dfb818c3f53f8eed0d5fca1e06bc1349c
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/267632
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Lin Jr, Chang-Sun <chang-sun.lin.jr@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>