[SYCL] Make host accessor commands bypass leaf limit (#2321)
Extend functionality of CircularBuffer class to allow for a different way of
processing of host-accessors leaves. From now on host-accessors won't be
affected by leaf-limiting and, hence, they won't be removed from leaves list.

The new LeavesCollection class stores two distinct data structures:

    CircularBuffer<Command *> for generic commands; list of pointers to
    EmptyCommand (which is the blocked node of host-accessor) and map of xrefs
    to the list.

Map of xrefs is merely a mapping of EmptyCommand's pointer to iterator in the
list. This mapping allows for quick lookup if the command is there (in the
list).

The introduced class also provides forward-iterator fetched with
begin()/cbegin() and end()/cend() methods along with ability to fetch all the
leaves stored as a vector of pointers to Command.

This patch also removes previously introduced workaroud (see commit a0199b3).

Change-Id: I51452ccb549e2a9e72ad3e6e3f92f3e1f7c0c524
Signed-off-by: Sergey Kanaev <sergey.kanaev@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/269105
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>