[Bitstream] Use alignTo to simplify some code and fix a 32-bit arithmetic wrap issue.
We need to convert to 64-bit before multiplying by 8 to avoid
overflow if NumElts is more than 512 megabytes. Using alignTo
will automatically do that and make the code more readable too.

We will still break again if the size of a blob exceeds 4GB, but
that requires changes to the reader and writer.

I plan to push this patch to community as well, but figured we
probably need it for gold.

Fixes CMPLRLLVM-20286

Change-Id: I0c66081eded44238c114cd739dff10d35e0b1c35
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/268519
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Reviewed-by: Sotkin, Alexey <alexey.sotkin@intel.com>