[traceback] Support debug information for multiple files case
Before this patch, we didn't handle the situations when there
are multiple files in one module, these cases could be

1. A compile unit includes a header file.
2. There are multiple compile units in one module.

Example 1:
```
// This is file test.c

int main() {
  return function_in_test_h();
}
```
The debug information for the source code in test.h was mapped to
test.c incorrectly.

Example 2:
The functions from test1.c, test2.c may be ovelapped when LTO is
enabled, we didn't consider this case too.

This patch fixes these issues.

Change-Id: Ic16aef4a7172d824e5b3d8f442db5b048be1b048
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/269275
Tested-by: Kan ShengChen <shengchen.kan@intel.com>
Reviewed-by: xiangzhang <xiang1.zhang@intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>