[NFCI][SYCL] Rework of DeclBodyCreator/Array Handling in prep of non-Decomposition (#2370)
The intent of this patch is simply to clean up how arrays work in the
SYCL visitors. The patch unifies the handling of array elements and
normal fields, such that they should be more maintainable.

This required a handful of changes to make integration headers work
a little better.

Work on this DID require quite a bit of a re-imagining of how the
DeclBodyCreator works. There are a number of small refactors to fix
things I found while working that I believe make it easier to maintain
this in the future.

However, the biggest change here is with the init-list-expr creation.
This patch creates them in advance, rather than waiting until we are
done with a struct/array/etc. This allows us to generate them as we go.

The side effect of this is that the InitExprs array now ONLY contains
collection InitListExprs, since it no longer has to assemble them on the
other side. However, now adding a field ends up being really easy,
since appending to the init-list is easier.

MemberExprBases also becomes easier to maintain, since they now only
get added when changing to/from a field or collection.

Change-Id: I385d94d319161b459256bd4a24ebdd95f00992ea
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/270930
Reviewed-by: Keane, Erich <erich.keane@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Andrews, Elizabeth <elizabeth.andrews@intel.com>
Reviewed-by: Huang, Zao <zao.huang@intel.com>