[InstCombine] Prevent combining two GEPs if one is just there to make the result type a vector.
This fixes a bug exposed by a recent change to
that replaces shuffle splats on scatters with a
GEP that uses a scalar base and all zero vector index.

The patch here specifically detects that sort of GEP,
but I'm not sure there aren't still other vector/scalar mismatch
issues in this combine.

Fixes CMPLRLLVM-22568 and CMPLRLLVM-22491

Change-Id: I84313da7853bcd0f8667357f01f61a83e8528640
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/268995
Reviewed-by: Topper, Craig <craig.topper@intel.com>
Tested-by: Topper, Craig <craig.topper@intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>