[BPI] Prioritize loop heuristic over cold call heuristic.
If a br instruction is post-dominated by a cold call, the same probability
value will be applied to all of its outgoing edges. This means the loop
branch heuristic won't function in this case, and loop body will have much
lower freqeuency that it needs:

 ...
  |
  v
 Body<---+
  |      |
  |      |
  v      |
Latch----+
  |   50%
  |50%
  v
ColdCall

However being post-dominated by a cold call doesn't mean the loop is not hot.
This can mislead CodeGen to generate inefficient code for the loop.

Change-Id: I061187d48720914700fd6906fa605b29cb6217e2
JIRA: CMPLRLLVM-12057
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/269637
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>
Tested-by: Tianqing Wang <tianqing.wang@intel.com>