[SATest] Fix flaky test_FP_CONTRACT_on.tst runfail
The device indicator in program configuration is uninitialized, so it
runs on EyeQ emulator sometimes, on which the FP_CONTRACT is disabled by
default.

Fix it by initialize the device indicator correctly, making the test
always run on CPU device. However, UnsafeSextInRegSimplification.tst
fails with FP_CONTRACT on. Disabling it temporarily.

Change-Id: I1cd2e391498fff17b9652e0c8648ed36b4bbc6a8
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/269034
Tested-by: Link Ren <link.ren@intel.com>
Reviewed-by: Sidorov, Dmitry <dmitry.sidorov@intel.com>