[Driver][SYCL] Improve tool name specification for AOT tools on Windows (#2389)
For Windows, we need to be sure to search for the full executable name
as opposed to just the case name for the AOT tools (aoc, ocloc, opencl-aot).
There can be instances of directory names of the same name, causing the
wrong 'binary' to be called.

Change-Id: I99e704a5f1de39d68f23a1cdab88b8c9c7852be3

[Driver][SYCL] Improve situations where .exe is added for AOT tools (#2467)

Previous fix only allowed for the .exe to be added for 'clang-cl' specific
compilations, which did not allow for usage using clang with a windows
specific target.

Addresses: https://jira.devtools.intel.com/browse/CMPLRLLVM-21438
cherry-pick of: https://github.com/intel/llvm/commit/244e874bbea9bee1ccdd7a08d0e7ba0218c2fda1

Change-Id: Ic5e1a4d329a12c40029263517e360327f7a6e93f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/272203
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Chilamantula, Harini <harini.chilamantula@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/272465
Tested-by: Toguchi, Michael D <michael.d.toguchi@intel.com>
Reviewed-by: Dale, Stewart T <stewart.t.dale@intel.com>