[DTrans - OpaquePtr] Fix null-ptr dereference for DTransSafetyAnalyzer
When a store instruction was used to store a pointer to structure type
into a pointer-to-pointer location for the structure type there was
the possibility of a crash from dereferencing a null pointer. Fix the
problem by using a variable that already holds the expected type of
the store instruction's value operand, which has already been checked
as being non-null.

Add test case for the case that was crashing, and another case where the
variable would be nullptr.

Change-Id: I9e77d5e90f0e40bec3f6d8ff442b588f478fab9f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/269438
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>