This is a cherry pick from xmain.
https://git-amr-2.devtools.intel.com/gerrit/#/c/266098/

The MS ABI requires the instantiation of defaut arguments for
default constructor for the purpose of generating constructor
closures. This patch does that.
This patch has been presented to community, but didn't get
any tracking. This crash revealed itself in the dpcpp build
(was in the MFL but has been pushed at a later date). This
solution is offered under a customization but might be open
sourced later.
Fix for https://jira.devtools.intel.com/browse/CMPLRLLVM-18148
The community patch is here:
https://reviews.llvm.org/D79591

Change-Id: I89c690172c164a9040be74f2fa305970809843db
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/269934
Reviewed-by: Ammarguellat, Zahira <zahira.ammarguellat@intel.com>
Tested-by: Ammarguellat, Zahira <zahira.ammarguellat@intel.com>