[SYCL] Diagnose attempt to pass pointer to VLA as kernel arg (#2441)
VLA as well as pointers to VLA require additional AST transformation to make
emission of corresponding type in LLVM IR possible. Without this transformation
Codegen just crashes.
Implementing this transformation is not reasonable whereas SYCL standard
doesn't allow VLA in device code, so emit an error in case if pointer to
VLA is passed as kernel argument.

Change-Id: I4eec7c9383066b55b7131e5bc3b5dfc212d938a7
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/270934
Reviewed-by: Andrews, Elizabeth <elizabeth.andrews@intel.com>
Tested-by: Andrews, Elizabeth <elizabeth.andrews@intel.com>