[SYCL] Improve checks of kernel execution range in INT_MAX limit. (#2423)
Improved check is for the following use-cases. When we have a lot of small work-groups the number of work-items may still be out of INT_MAX limit Also, sum of range and offset may exceed the limitation while each value is within INT_MAX limit on its own.

Signed-off-by: Sergey Kanaev <sergey.kanaev@intel.com>

This is cherry-pick of 57313f4f4c1266a79fb99c95933994dac298870b from syclos.
This patch is for CMPLRLLVM-22519.

Change-Id: Ieb0d680e3cb4b31d42a3fd1879fea09f9c6e453e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/270138
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Alexander Batashev <alexander.batashev@intel.com>
Reviewed-by: Sidorov, Dmitry <dmitry.sidorov@intel.com>