[SYCL] Improve checks of kernel execution range in INT_MAX limit. (#2423)
Improved check is for the following use-cases. When we have a lot of small work-groups the number of work-items may still be out of INT_MAX limit Also, sum of range and offset may exceed the limitation while each value is within INT_MAX limit on its own.

Signed-off-by: Sergey Kanaev <sergey.kanaev@intel.com>

This is cherry-pick of 57313f4f4c1266a79fb99c95933994dac298870b from syclos.
This patch is for CMPLRLLVM-22519.

Change-Id: I2932b9bf4c9c048bb2b65c3f3dab045f0f516df7
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/270140
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Alexander Batashev <alexander.batashev@intel.com>
Reviewed-by: Saito, Hideki <hideki.saito@intel.com>
Reviewed-by: Chupin, Pavel V <pavel.v.chupin@intel.com>
Reviewed-by: Jayashankar, Prema <prema.jayashankar@intel.com>