[VPlan][Predicator] Optimize blends with undef
Most common source of such undefs is loop exits canonicalization where we're
forced to introduce them to maintain SSA form and the earliest place in the
pipeline where we can get rid of them without breaking SSA is after
linearization.

The problem with them is that DA marks phis like [uniform, undef] as uniform and
that puts restrictions on what later phases are allowed to do. Treating them as
divergent would result in bad performance sometimes and properly marking them as
uniform needs either undef elimination or active lane extraction. The latter is
bad for perf as well, so just optimize undef out.

Part of the fix for CMPLRLLVM-21787 - to avoid performance regressions
happenning otherwise.

Change-Id: I65d0966d3132b0a156085f7b4350121306dac63d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/270489
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>