[VPlan][Predicator] Optimize out non-LCSSA single incomving value phis
Historically, we were converting such phis into blends to avoid emitting phis in
the middle of the llvm::BasicBlock due to the reuse of the predecessor
VPBasicBlock inside VPlan LLVM IR CG. However, just optimizing such phis away
seems to be clearer and simpler.

Part of the fix for CMPLRLLVM-21787.

Change-Id: Ib238f6bd8393072d6d098dcdb676ddeb582304cd
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/270490
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>