[NFC][loopopt] Switch distance/direction vector representations to SmallVector
Previously these vectors were represented as std::array with a fixed
maximum size, where unused elements were indicated with a particular
value. ("NONE", with value 0.) For example, a direction vector with
length 3 was represented as:

    (* * = 0 0 0 0 0 0)

Recent improvements to dependence testing allow inner loops to be
determined to have no loop-carried or loop independent dependence, which
is also represented with a "NONE"/0 value. This results in some
ambiguity, essentially leaving the two direction vectors below
impossible to distinguish between without additional context:

    (* *)
    (* * 0)

To make the representation more clear, the vectors are now represented
as LLVM SmallVector containers. This allows the intended vector length
to be stored explicitly rather implicitly based on the number of
rightmost zero elements.

The benefit of this change can be seen in the scoped-aa-varying.ll LIT
test it updates, where distance vectors for a two-level loop nest can
now be printed as:

    (%B)[i2 + -1] --> (%A)[i2] ANTI (* 0)
    (%A)[i2] --> (%B)[i2 + -1] FLOW (* 0)

...rather than:

    (%B)[i2 + -1] --> (%A)[i2] ANTI (*)
    (%A)[i2] --> (%B)[i2 + -1] FLOW (*)

Change-Id: I72da4ee0e1a86be4d95bae9b55d8a8e9741ad581
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/270758
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>