[CMPLRLLVM-22495] : coremark-pro/zip-test@opt_base_SKL(32-bits) has -6% regression after community pulldown August 6-August 24, 2020
The direct cause is from D32255 that transfers "xor a, 65535" to "xor a, -1".
But the root cause is register allocation’s limitation that it
doesn’t has the ability to think about global optimum, so that each
current highest weight variable always occupies a register by evicting
other variable, and it possible leads to a series variables’
spilling and reloading.

Since it is hard to fix the regression in current RA algorithm and framework,
I fix it by modifying branch probability so that the weight in RA is a little
bit more accurate than general.

I add a new function to judge a Likely branch in loop, and make its weight a
little bigger than general edge. (InEdge is 124, LikelyEdge is 135. And the 135
is the minimum value which could work for this case) for below case:
int n = a;
while (...) {
  if (++n < invariant) {
    do sth;
  }
}

If need, we could extense the pattern and adjust weight value in the future for
other cases performance.

Change-Id: Ibbc62684dea12d5c40d22492110e710ee2cae622
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/270622
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Lin Jr, Chang-Sun <chang-sun.lin.jr@intel.com>
Reviewed-by: Bakhvalov, Denis <denis.bakhvalov@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Sotkin, Alexey <alexey.sotkin@intel.com>