[NFC][VPlan] Move the checks that determine interleaving to a separate function
Changes here move the various checks that are done to see if we should
interleave a memory access to a separate function. The checks are also
modified to be VPInstruction based. This is done so that we can use this
function in an upcoming change set that enables VLS interleaving in
VPValue based code generation.

Change-Id: I053a63e62412c084bd4b4b4576bebf59ba51fdf8
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/270708
Reviewed-by: Pushkar Ratnalikar <pushkar.v.ratnalikar@intel.com>
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Sotkin, Alexey <alexey.sotkin@intel.com>