[NFC][VPlan] Simplify getMemoryRef interface to take fewer arguments
We were computing and passing several arguments to getMemoryRef interface
from a VPLoadStoreInst instruction. The change here passes VPLoadStoreInst
to getMemoryRef which now computes the needed values. The change is being
made in preparation to support VLS interleaving of memory accesses in
VPValue based code generation which is expected add uses of getMemoryRef
interface.

Change-Id: I8ba68ae5851e638d11d32ea5d515120e433ddbfc
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/270709
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: Guggilla, Satish <satish.guggilla@intel.com>