[SYCL] Support or diagnose use of namespace std types as kernel type names (#1963)
When std::nullptr_t is used as a kernel type, the generated
integration header uses 'nullptr_t'.  This causes lookup
errors.  Use 'std::nullptr_t' instead.

std::max_align_t is defined (in one implementation) as a typedef
of an anonymous struct.  This causes errors when attempting to
forward declare the type in the integration header.  Diagnose
such cases earlier.

Change-Id: Ibfc8eb8cf6b5586004f6a8a0f1fa6ebbda70f9ff
Signed-off-by: Premanand M Rao <premanand.m.rao@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/278628
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Huang, Zao <zao.huang@intel.com>
Reviewed-by: Wang, Jing W <jing.w.wang@intel.com>