[CMPLRLLVM-22598] [Andersens] Fix target function selection for indirect calls check of number of arguments
In the original problem case that resulted in a failure, the source code was
casting a function pointer type which took 4 arguments into a function
pointer type that takes 2 arguments, and then making an indirect function
call. This is likely incorrect source code, but should not cause the
compiler to crash.

When the Indirect call conversion pass requested a list of
possible call targets from Andersen for the now 2 argument callsite,
a function that required 4 arguments was returned, and a direct function
call to the function was made. Later, when the inliner tried to inline
the direct function call, it failed with a bad memory reference when
looking for values for the additional function arguments needed by
the function.

Change-Id: Ic282724c699d1f1df1556e91096cba968b2f6a93
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/271243
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/271568
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Jayashankar, Prema <prema.jayashankar@intel.com>