* [Driver][SYCL] Make /MD the default for -fsycl (#2478)
When using -fsycl on Windows, make /MD the default behavior.  Any usage
of /MT will not be allowed and the driver will error upon usage.

cherry pick of: https://github.com/intel/llvm/commit/d31184e11fccae0b241df9c85e136f23dc2ad73c

Change-Id: Ib275170612c4782239074f2f015dacb467501e20
Addresses: CMPLRLLVM-13952
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/271334
Reviewed-by: Toguchi, Michael D <michael.d.toguchi@intel.com>
Tested-by: Toguchi, Michael D <michael.d.toguchi@intel.com>

* [SYCL][Driver] Make /MD default with -fsycl for clang/clang++ drivers

SYCL library is designed such a way that STL objects must cross the sycl.dll
boundary, which is guaranteed to work safe on Windows only if the runtime
in the app using sycl.dll and in sycl.dll is the same and is dynamic.

It is not possible to implement safe approach for using sycl libraries
built/linked with static C++ RT as it would cause having multiple copies
of C++ objects (such as scheduler, etc), which are supposed to be singletones.

sycl.dll is built with /MD (linked to dynamic C++ RT), this is why this patch
makes /MD default with -fsycl for clang and clang++ drivers.

Change-Id: I46436ffde7dab4d3afc9486b6948b423923a592e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/271916
Reviewed-by: Klochkov, Vyacheslav N <vyacheslav.n.klochkov@intel.com>
Tested-by: Klochkov, Vyacheslav N <vyacheslav.n.klochkov@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>

* [Driver][SYCL] Adjust /MD settings for SYCL device compilations (#2497)

When setting /MD (by default) we do not want to pass the _MT and _DLL
predefines to the device compilations, only the host for clang-cl

Taken from: https://github.com/intel/llvm/commit/0092d4da23a16712e1cb15bee067cb3df3565955
Addresses: CMPLRLLVM-13952

Change-Id: I141205c766aff58aa386f1a954e1af7b5aa7c6cb
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/272258
Reviewed-by: Klochkov, Vyacheslav N <vyacheslav.n.klochkov@intel.com>
Tested-by: Klochkov, Vyacheslav N <vyacheslav.n.klochkov@intel.com>