[VPlan] Add a test showing bug in recomputation of shapes of instructions dependent on SOA-privates.
We currently compute the shape of SOA-private alloca's and then push the
dependent instructions to the WL for iteratively updating all users. We,
however, exclude instructions which are outside the loop. This results
in an issue where instruction like GEP, which can be in the
loop-preheader would not be updated with SOA-shape resulting in
incorrect information for the CG to work on.

Change-Id: Ic56c99218a81e82bee260768cb5c49af909f6bba
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/271450
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Pushkar Ratnalikar <pushkar.v.ratnalikar@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>