Manually cherry-pick two commits from xmain because they have dependency: 1. https://git-amr-2.devtools.intel.com/gerrit/#/c/270056/ [CMPLRLLVM-22704] : Fix CurrentDT assert issue.
2. https://git-amr-2.devtools.intel.com/gerrit/#/c/270622/
[CMPLRLLVM-22495] : coremark-pro/zip-test@opt_base_SKL(32-bits)
has -6% regression after community pulldown August 6-August 24, 2020

The direct cause is from D32255 that transfers "xor a, 65535" to "xor a, -1".
But the root cause is register allocation’s limitation that it
doesn’t has the ability to think about global optimum, so that each
current highest weight variable always occupies a register by evicting
other variable, and it possible leads to a series variables’
spilling and reloading.

Since it is hard to fix the regression in current RA algorithm and framework,
I fix it by modifying branch probability so that the weight in RA is a little
bit more accurate than general.

I add a new function to judge a Likely branch in loop, and make its weight a
little bigger than general edge. (InEdge is 124, LikelyEdge is 135. And the 135
is the minimum value which could work for this case) for below case:
int n = a;
while (...) {
  if (++n < invariant) {
    do sth;
  }
}

If need, we could extense the pattern and adjust weight value in the future for
other cases performance.

Change-Id: I41cd5034aedd84aadead2fadb78808a466c0282d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/271676
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Xiao, Wei3 <wei3.xiao@intel.com>
Reviewed-by: Huang, Zao <zao.huang@intel.com>