[Paropt] Add range metadata to signed LB/UB for parallel loops
Parallel loops are normalized by the front end, so outlined loop's LB/UB
are expected to be positive if IV is signed. This patch adds range
metadata to the LB/UB loads indicating that LB/UB values are in range
[0, MaxValue), where MaxValue is either the parallel loop's TC+1 if loop
bounds are known or the maximum signed value of the IV's type otherwise.

Change-Id: If77b14256abee1d1af831805cb147aac511c2c8b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/271490
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Pengfei Wang <pengfei.wang@intel.com>