[VPlan] Adding linear-index property to min/max+index idiom
This patch adds possibility to have non-inductive values in min/max+index idioms.
E.g. the following idioms can be supported:
  if (b[i] < a) {
     a = b[i];
     ndx = i;
     t = c[i];  // this is possible to calculate now correctly
  }
The "ndx" here is marked as a linear-index and is set as parent for "t"-reduction
while "a" is set as parent for "ndx".
Along with data structure, the code generation is added. The following code is
generated for index parts of min/max+index idiom.
Here parent_final is last value of the parent reduction, parent_exit is final
vector of the parent reduciton, reduce_val is the final vector of the current
reduction.
For IsLinearIndex == true (for the example above, the code for ndx's last value):
  %bcst = broadcast parent_final  // parent here is main reduction
  %cmp_v = cmp eq %bcst, parent_exit
  %ndx_fixup = select %cmp_v  %reduce_val, (is_min(pred) ? MAX_INT:MIN_INT)
  %result = (is_min(pred) ? reduce_min: reduce_max)(%ndx_fixup)

For IsLinearIndex == false (for the example above, the code for t's last value):
  %bcst = broadcast parent_final // parent here is a linear index
  %cmp_v = cmp eq %bcst, parent_exit
  %ndx = cttz(cmp_v)
  %result =  extract_element reduce_val, %ndx
}

The floating point data type for main reduction is not supported yet.

Change-Id: I2f7ff40aa30fd856af6be403b15af17c7250ccf3
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/209572
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Tested-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>