[VPlan] Implement VLS support in VPValue based code generation for HIR
Changes in this patchset add support for generating interleaved loads/stores
in VPValue based code generation for HIR. Changes also implement propagation
of metadata from instructions in VLS group to the newly generated memory
references for the interleaved accesses.

JIRA : CMPLRLLVM-21142

Changes here combine the following xmain changes that were added for
implementing this support and fix the 29% performance regression in
specACCELref/514@opt_base6_hsw_ihost.

[NFC][VPlan] Move the checks that determine interleaving to a separate function
    --- https://git-amr-2.devtools.intel.com/gerrit/#/c/270708/
[NFC][VPlan] Simplify getMemoryRef interface to take fewer arguments
    --- https://git-amr-2.devtools.intel.com/gerrit/#/c/270709/
[NFC][VPlan] Templatize propagateMetadata interface to work with RegDDRefs/VPLoadStoreInsts
    --- https://git-amr-2.devtools.intel.com/gerrit/#/c/270710/
[VPlan] Implement VLS support in VPValue based code generation for HIR
    --- https://git-amr-2.devtools.intel.com/gerrit/#/c/270781/

Cherry picking these changes into xmain-rel did not work and I had to
resort to manually applying the files. The only difference in the applied
changes is due to support of stores to a uniform location which is only
implemented in xmain and was not allowed to be applied to xmain-rel.

Change-Id: I67cdd2aafff53ee0d453c031fc9753823e87f78c
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/271881
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Prokofiev, Pavel G <pavel.g.prokofiev@intel.com>
Reviewed-by: Jayashankar, Prema <prema.jayashankar@intel.com>