[VPlan] Fix handling of casts and lifetime_start/end intrinsics in the CodeGen.
This patch generates efficient code for the lifetime_start/end
intrinsics. Currently, we serialize the call to these intrinsics and
generate VF-calls, passing the base-address of each private-copy to each
of those calls. This is necessary only when we serialize the widening of
the alloca's. If we generate a VF-wide alloca, only one call to these
intrinsics is necessary.

This patch also fixes the size computation, i.e., the first argument
that is to be passes to these intrinsics.

In addition, a couple of helpers are added to the VPCallInstruction
class and corresponding refactoring is done.

Change-Id: I10f9585289002e44ec488229abe4f6acddf325c3
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/271987
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Pengfei Wang <pengfei.wang@intel.com>