[SYCL][Driver] Make /MD default with -fsycl for clang/clang++ drivers
SYCL library is designed such a way that STL objects must cross the sycl.dll
boundary, which is guaranteed to work safe on Windows only if the runtime
in the app using sycl.dll and in sycl.dll is the same and is dynamic.

It is not possible to implement safe approach for using sycl libraries
built/linked with static C++ RT as it would cause having multiple copies
of C++ objects (such as scheduler, etc), which are supposed to be singletones.

sycl.dll is built with /MD (linked to dynamic C++ RT), this is why this patch
makes /MD default with -fsycl for clang and clang++ drivers.

Change-Id: I46436ffde7dab4d3afc9486b6948b423923a592e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/271916
Reviewed-by: Klochkov, Vyacheslav N <vyacheslav.n.klochkov@intel.com>
Tested-by: Klochkov, Vyacheslav N <vyacheslav.n.klochkov@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>