[CMPLRLLVM-23149] [DTrans] Fix for DTransSafety bits incorrectly set, allowing deletefield to try to modify a structure
The code introduced to the DTransAnalysis to inhibit the safety checks needed
to implement the EliminateROFieldAccess transformation marked memory allocations
as safe that matched the pattern in that program without taking into account
that other transformations, such as deletefield, are unable to transform those
allocations.

This change adds the "Complex alloc size" safety bit to identify allocations
that cannot be easily have their allocation sizes transformed, but are still
valid for the purpose of identifying values written to fields of the structure.

Change-Id: I01ec4f30a75293f5bf539ca5a4f25e7bf725adcf
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/272434
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>