[SYCL] Fixed bugs regarding platform caching
Added new platforms that have been created and initialized to the PiPlatformCache.

Created a new function, getOrCreatePlatform, so that the cached platforms could
be shared across both piPlatformsGet and piextPlatformCreateWithNativeHandle.

Handle platform initialization in piCachePlatforms, so that platforms created
with a native handle would also be initialized (before initialization was not
occurring for platforms created in piextPlatformCreateWithNativeHandle).

Change-Id: I81e5742af696fea07cc0c9df34c32a96a6653a93
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/272365
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Tested-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Reviewed-by: Huang, Zao <zao.huang@intel.com>