CMPLRLLVM-23200: Fix regression where pointers in functor base crashes
This regression was introduced when the large refactor happened, so it
is in xmain-rel at the moment, and thus needs fixing.  It is fixed in a
syclos patch, but that is likely too high risk to go to xmain-rel, so we
are fixing this separately.

Basically, pointers-in-functor-bases weren't being wrapped properly
before the refactor patch (as they weren't being wrapped). However, the
refactor only accidentially implemented 1/2 of the wrapping behavior.

This patch adds the other half, plus a test that crashes without this
fix.

Change-Id: Ic48d0207e3ef40575a45dba3fe117975281b3991
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/272650
Reviewed-by: Keane, Erich <erich.keane@intel.com>
Tested-by: Keane, Erich <erich.keane@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>