[VPlan] Add the '-vplan-enable-soa=false' option to all relevant  tests in preparation of committing of the SOA-CodeGen patch.
These tests are updated to explicitly disable SOA codegen and retain the
existing behavior. When the SOA CG changes are pushed and feature
switched on by default, it is likely to cause failures for the following tests.

This patch also has some refactor of the SOA-analysis switch, moving the
switch to VPlan-class and explicitly disable it for HIR path.

Change-Id: I9ce2df1c8595ef7ef4d7394f2ded2adaf8b6c814
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/272271
Reviewed-by: Pushkar Ratnalikar <pushkar.v.ratnalikar@intel.com>
Tested-by: Pushkar Ratnalikar <pushkar.v.ratnalikar@intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>