Introduce intel.honor.fcmp intrinsic
In the IL0 compilers, 'fp-model fast' allowed optimizations that assumed
no NaN inputs or outputs, but it had a semantic mode 'ieee_nan_compares'
which disallowed that assumption for fcmp instructions.

Simply removing the 'nnan' flag from fcmp instructions in LLVM isn't
sufficient because if the flag is set for the instructions that provide
the operand values the LLVM optimizer will assume the inputs are not NaN
even if the nnan flag isn't set on the fcmp instruction.

To achieve the equivalent behavior of the IL0 compiler, we need to
introduce an intrinsic that imposes different semantics on the fcmp
instruction.

By default, we will use the clang behavior for fp-model fast in the
LLVM-based compilers, but we want to add a user option to get icc/ifort
compatible behavior.

This change adds only the LLVM support for the new intrinsic. Clang and
Fortran front end changes will be added in separate patches.

Change-Id: I88819481558f495be216c01f6617a29721fea537
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/273242
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Tested-by: Kaylor, Andrew <andrew.kaylor@intel.com>