Add clang front end support for -fhonor-nan-compares
In the IL0 compilers, 'fp-model fast' allowed optimizations that assumed
no NaN inputs or outputs, but it had a semantic mode 'ieee_nan_compares'
which disallowed that assumption for fcmp instructions.

To achieve the equivalent behavior of the IL0 compiler, we are
introducing a new command line option (-fhonor-nan-compares) that causes
the front end to avoid setting the 'nnan' fast-math flag on any
instructions.

By default, we will use the clang behavior for fp-model fast in the
LLVM-based compilers, but we want to add a user option to get icc/ifort
compatible behavior.

Change-Id: Id91a33a6405b0e6b66ecbfe8209fa17a20b2a7b8
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/276299
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Tested-by: Kaylor, Andrew <andrew.kaylor@intel.com>