[VPlan][Minmax+index] Correction to create VPExternalUse for main reduction in minmax+index idiom.
We don't create VPExternalUse for reductions which are not used after loop.
In HIR path, in case when we have minmax+index idiom and the main reduction is not used
after loop, we put its last value to a incorrect temp. Then calculations in the
remainder loop start from the incorrect value.
The fix is to create VPExternalUse with a correct DDREF, which is taken from
incoming value.

Change-Id: I6ee9c9579c1dcf040bf3308aa8cd2fcfc536e73b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/273250
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Pengfei Wang <pengfei.wang@intel.com>