[LLD] CMPLRLLVM-23292: Remove cl::ResetAllOptionOccurrences from the ELF driver
This change set comments the call to cl::ResetAllOptionOccurrences in the LLD
ELF driver. This function resets the clang options, but it is called inside a
for loop. It wipes the options collected. For example, if the user adds
multiple -mllvm options, then the last -mllvm will be collected.

This function was added in the following change set by the community:
https://reviews.llvm.org/D70378

********************
Testing:  0.. 10.. 20.. 30.. 40.. 50.. 60.. 70.. 80.. 90..
********************
Failed Tests (1):
  LLVM-Unit :: Support/./SupportTests/CrashRecoveryTest.UnixCRCReturnCode

Testing Time: 1448.73s
  Unsupported      : 19168
  Passed           : 56188
  Expectedly Failed:   192
  Failed           :     1

Change-Id: I7d3da6b9b370d1be6a112f32a2795e43b267027a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/273277
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: Cox, Robert <robert.cox@intel.com>